Document the behavior when creating a new env using an existing env name
Describe an un-common behavior/feature when creating an environment when an existing environment name is used. In this case, instead of returning 400 or 409 as normal RESTful service does, the call/operation is successful with a new environment created. The new environment has a unique ID, but the name is not the name you entered, but a "-<n>" appended to it, the <n> is the incremental number. This is working as designed as per the conversation here: https://mulesoft.slack.com/archives/G7JFWLX54/p1598918390006400